Footer styles created
